agent/requirements.txt
websockets>=12.0
pydantic>=2.0
pydantic-settings>=2.0
PyYAML>=6.0
